+++ source/regchecklistbox.pas
@@ -26,7 +26,6 @@ type
     FRegistrySettings: TRegistrySettingsCheckedList;
     FIsModified: boolean;
 
-    function GetKeyByItemIndex(aItemIndex: integer): string;
     function GetKeyBySourceKind(aItemIndex: integer): string;
     procedure SaveItemsStatesToReg(aOnlyByItemIndex: boolean = True);
     procedure SetCheckedItemsByList(aList: TStrings;
@@ -70,7 +69,9 @@ type
       read FRegistrySource
       write SetRegistrySource;
   public
-    procedure DeleteItem(aItemIndex: integer = -1);
+    function DeleteItem(aItemIndex: integer = -1;
+                        aAskFor: boolean = True;
+                        aMsg: string = 'Delete Item?'): boolean; virtual;
     function ClearItems(aAskFor: boolean = True;
                         aMsg: string = 'Clear Items?'): boolean; virtual;
     procedure Click; override;
@@ -115,56 +116,11 @@ end;
 
 { TCustomRegCheckListBox }
 
-function TCustomRegCheckListBox.GetKeyByItemIndex(aItemIndex: integer): string;
-var
-  list: TStrings;
-begin
-  Result := EmptyStr;
-  list := TStringList.Create;
-  try
-    try
-      if not (csDesigning in ComponentState) then
-      begin
-        if assigned(RegistrySource) then
-        begin
-          if ((FRegistrySettings.RootKey <> '') and
-            (FRegistrySettings.RootKeyForDefaults <> '') and
-            (FRegistrySettings.RootForDefaults <> '') and
-            (FRegistrySettings.ListSection <> '')) then
-          begin
-            RegistrySource.ReadSection(FRegistrySettings.RootKey,
-              FRegistrySettings.RootKeyForDefaults,
-              FRegistrySettings.RootForDefaults,
-              FRegistrySettings.ListSection,
-              list,
-              FRegistrySettings.DoMergeData,
-              FRegistrySettings.ReadDefaults,
-              lskByKeyValue);
-
-            if (aItemIndex <= list.Count-1) then
-              Result := list.Names[aItemIndex];
-          end;
-        end;
-      end;
-    except
-      on E: Exception do
-        Result := EmptyStr;
-    end;
-  finally
-    if Assigned(list) then
-      FreeAndNil(list);
-  end;
-end;
-
 function TCustomRegCheckListBox.GetKeyBySourceKind(aItemIndex: integer): string;
 begin
   case RegistrySettings.SourceKind of
-    lskByKey:
-      Result := Items.Strings[aItemIndex];
-    lskByValue:
-      Result := GetKeyByItemIndex(aItemIndex);
     lskByKeyValue:
-      Result := Items.Names[aItemIndex];
+      Result := Items.Strings[aItemIndex];
   else
     MessageDlg('Invalid SourceKind!',
       mtWarning, [mbOK], 0);
@@ -623,9 +579,12 @@ begin
   end;
 end;
 
-procedure TCustomRegCheckListBox.DeleteItem(aItemIndex: integer = -1);
+function TCustomRegCheckListBox.DeleteItem(aItemIndex: integer = -1;
+  aAskFor: boolean = True;
+  aMsg: string = 'Delete Item?'): boolean;
 var
   key: string;
+  start: boolean;
 begin
   if (aItemIndex = -1) then
     aItemIndex := Self.ItemIndex
@@ -633,20 +592,32 @@ begin
   if ((aItemIndex < 0) or (aItemIndex > Items.Count-1)) then
     raise Exception.CreateFmt('Invalid Index: %d', [aItemIndex]);
 
-  key := GetKeyBySourceKind(aItemIndex);
-  if (key <> EmptyStr) then
-    if ((FRegistrySettings.RootKey <> '') and
-        (FRegistrySettings.RootKeyForDefaults <> '') and
-        (FRegistrySettings.RootForDefaults <> '') and
-        (FRegistrySettings.ListSection <> ''))
-    then
-      FRegistrySource.DeleteKey(FRegistrySettings.RootKey,
-        FRegistrySettings.RootKeyForDefaults,
-        FRegistrySettings.RootForDefaults,
-        FRegistrySettings.ListSection,
-        key,
-        FRegistrySettings.WriteDefaults,
-        FRegistrySettings.GroupIndex);
+  start := not aAskFor;
+
+  if FRegistrySettings.ItemsByRegistry then
+  begin
+    if aAskFor then
+      start := (MessageDlg(aMsg, mtConfirmation, [mbYes, mbNo], 0) = mrYes);
+    if start then
+    begin
+      key := GetKeyBySourceKind(aItemIndex);
+      if (key <> EmptyStr) then
+        if ((FRegistrySettings.RootKey <> '') and
+            (FRegistrySettings.RootKeyForDefaults <> '') and
+            (FRegistrySettings.RootForDefaults <> '') and
+            (FRegistrySettings.ListSection <> ''))
+        then
+          FRegistrySource.DeleteKey(FRegistrySettings.RootKey,
+            FRegistrySettings.RootKeyForDefaults,
+            FRegistrySettings.RootForDefaults,
+            FRegistrySettings.ListSection,
+            key,
+            FRegistrySettings.WriteDefaults,
+            FRegistrySettings.GroupIndex);
+    end;
+  end;
+
+  Result := (start and FRegistrySettings.ItemsByRegistry);
 end;
 
 function TCustomRegCheckListBox.ClearItems(aAskFor: boolean = True;
